refactor: 🧪 DeepReadonly
