+++ test/index.ts
@@ -271,86 +271,6 @@ function testDeepNullable() {
   ];
 }
 
-function testDeepReadonly() {
-  type cases = [
-    Assert<IsExact<DeepReadonly<number>, number>>,
-    Assert<IsExact<DeepReadonly<string>, string>>,
-    Assert<IsExact<DeepReadonly<boolean>, boolean>>,
-    Assert<IsExact<DeepReadonly<bigint>, bigint>>,
-    Assert<IsExact<DeepReadonly<symbol>, symbol>>,
-    Assert<IsExact<DeepReadonly<undefined>, undefined>>,
-    Assert<IsExact<DeepReadonly<null>, null>>,
-    Assert<IsExact<DeepReadonly<Function>, Function>>,
-    Assert<IsExact<DeepReadonly<Date>, Date>>,
-    Assert<IsExact<DeepReadonly<Error>, Error>>,
-    Assert<IsExact<DeepReadonly<RegExp>, RegExp>>,
-    Assert<IsExact<DeepReadonly<Map<string, boolean>>, ReadonlyMap<string, boolean>>>,
-    Assert<IsExact<DeepReadonly<ReadonlyMap<string, boolean>>, ReadonlyMap<string, boolean>>>,
-    Assert<IsExact<DeepReadonly<WeakMap<{ key: string }, boolean>>, WeakMap<{ key: string }, boolean>>>,
-    Assert<
-      IsExact<DeepReadonly<WeakMap<{ key: string }, { value: boolean }>>, WeakMap<{ key: string }, { value: boolean }>>
-    >,
-    Assert<IsExact<DeepReadonly<Set<string>>, ReadonlySet<string>>>,
-    Assert<IsExact<DeepReadonly<ReadonlySet<string>>, ReadonlySet<string>>>,
-    Assert<IsExact<DeepReadonly<[]>, readonly []>>,
-    Assert<IsExact<DeepReadonly<[1, 2, 3]>, readonly [1, 2, 3]>>,
-    Assert<IsExact<DeepReadonly<readonly number[]>, readonly number[]>>,
-    Assert<IsExact<DeepReadonly<Array<number>>, ReadonlyArray<number>>>,
-    Assert<IsExact<DeepReadonly<Promise<number>>, Promise<number>>>,
-    Assert<IsExact<DeepReadonly<{ a: 1; b: 2; c: 3 }>, { a: 1; b: 2; c: 3 }>>,
-    Assert<IsExact<DeepReadonly<{ foo: () => void }>, { foo: () => void }>>,
-    Assert<
-      IsExact<
-        DeepReadonly<{ obj: unknown; arr: unknown[] }>,
-        { readonly obj: unknown; readonly arr: readonly unknown[] }
-      >
-    >,
-    Assert<IsExact<DeepReadonly<ComplexNestedRequired>, ComplexNestedReadonly>>,
-  ];
-
-  // Build-time test to ensure the fix for
-  // https://github.com/ts-essentials/ts-essentials/pull/310
-  // because IsExact<> is unable to text it.
-  {
-    type TestUnion = { value: string } | TestUnion[];
-    type ReadonlyTestUnion = { readonly value: string } | readonly ReadonlyTestUnion[];
-
-    const a: DeepReadonly<TestUnion> = [];
-    const b: ReadonlyTestUnion = a;
-  }
-
-  // Build-time test to ensure the fix for
-  // https://github.com/krzkaczor/ts-essentials/issues/17 remains in place.
-  {
-    interface TestObject extends DeepReadonly<{ field: string[] }> {}
-
-    let a: DeepReadonly<TestObject> = {
-      field: ["lala"],
-    };
-
-    let b: TestObject = {
-      field: ["lala"],
-    };
-
-    b = a;
-  }
-
-  {
-    interface TestObject {
-      obj: unknown;
-      arr: unknown[];
-    }
-
-    // @ts-expect-error
-    const obj: TestObject = null;
-
-    let readonlyObj: DeepReadonly<TestObject>;
-    readonlyObj = obj;
-
-    const readonlyObj2: DeepReadonly<TestObject> = obj;
-  }
-}
-
 function testDeepUndefinable() {
   type cases = [
     Assert<IsExact<DeepUndefinable<number>, number | undefined>>,
